refactor(user-profile): extract estimator trace builders + core-math (WP2.1c)
Mechanical leaf extraction for Deep Refactor Plan v3 Phase 2. Cluster 2 trace
builders move to utils/_profile_estimator/traces.py and the core-math
primitives move to utils/_profile_estimator/core_math.py. utils/profile_estimator.py
stays the stable public facade/orchestrator and re-exports every moved name.

Moved verbatim (byte-identical vs HEAD, AST-verified): 10 trace builders
(_default, _format_experience_label, _format_rounding_label, _build_default_trace,
_build_log_trace, _build_profile_trace, _build_profile_bodyweight_trace,
_build_cold_start_trace, _build_learned_trace, _build_related_learned_trace) and
8 core-math names (_normalize_for_matching, _COMPLEX_ALLOWLIST_NORMALIZED,
classify_tier, epley_1rm, round_weight, _load_basis_factor,
_classify_experience_tier, cold_start_1rm). No renaming, logic consolidation,
typing expansion, or reordering; signatures, annotations, rounding, trace dicts,
exception behavior, and calculation order preserved.

Architecture: constants.py stays a pure leaf; traces.py and core_math.py import
only constants + utils.normalization and are mutually independent siblings.
Neither leaf imports strength_calibration — the profile_estimator <-> strength_
calibration cycle stays held solely by the two function-local lazy imports in the
facade orchestration (_lookup_related_learned_calibration, _lookup_learned_
calibration), so both import orders still resolve. Re-exports are direct
(# noqa: F401), preserving object identity (facade.epley_1rm is
strength_calibration.epley_1rm).

Verification: estimator contract 6 passed (surface + identity + both import
orders); focused estimator/calibration/user-profile suites 206 passed; full
pytest green (1703 + 13 catalog-dependent); flake8 blocking set 0; pyright
baseline gate PASS (0 net-new, no baseline regeneration needed — the moved
functions' reportArgumentType diagnostics anchor to their call sites, which
stay in the facade).

Migration: none; behavior and public imports preserved.
